Move SearchApiRoutes and SearchRestService to own files
