+++ webapi/src/main/scala/org/knora/webapi/slice/search/api/SearchEndpoints.scala
@@ -8,21 +8,14 @@ package org.knora.webapi.slice.search.api
 import eu.timepit.refined.api.Refined
 import eu.timepit.refined.api.RefinedTypeOps
 import eu.timepit.refined.numeric.Greater
-import io.opentelemetry.api.common.Attributes
-import io.sentry.Sentry
-import io.sentry.SentryLevel
-import org.apache.pekko.http.scaladsl.server.Route
 import sttp.model.HeaderNames
 import sttp.model.MediaType
 import sttp.tapir.*
 import sttp.tapir.codec.refined.*
-import zio.Task
 import zio.ZIO
 import zio.ZLayer
-import zio.telemetry.opentelemetry.tracing.Tracing
 
 import dsp.valueobjects.Iri
-import org.knora.webapi.responders.v2.SearchResponderV2
 import org.knora.webapi.slice.admin.api.Codecs.TapirCodec.*
 import org.knora.webapi.slice.admin.domain.model.KnoraProject.ProjectIri
 import org.knora.webapi.slice.admin.domain.model.User
@@ -30,10 +23,7 @@ import org.knora.webapi.slice.common.StringValueCompanion
 import org.knora.webapi.slice.common.Value.StringValue
 import org.knora.webapi.slice.common.api.*
 import org.knora.webapi.slice.common.api.KnoraResponseRenderer.FormatOptions
-import org.knora.webapi.slice.common.api.KnoraResponseRenderer.RenderedResponse
-import org.knora.webapi.slice.common.service.IriConverter
 import org.knora.webapi.slice.search.api.SearchEndpointsInputs.InputIri
-import org.knora.webapi.slice.search.api.SearchEndpointsInputs.Offset
 
 object SearchEndpointsInputs {
 
@@ -216,330 +206,4 @@ final case class SearchEndpoints(baseEndpoints: BaseEndpoints) {
 
 object SearchEndpoints {
   val layer = ZLayer.derive[SearchEndpoints]
-}
-
-final case class SearchApiRoutes(
-  searchEndpoints: SearchEndpoints,
-  searchRestService: SearchRestService,
-  mapper: HandlerMapper,
-  tapirToPekko: TapirToPekkoInterpreter,
-  iriConverter: IriConverter,
-) {
-  private type GravsearchQuery = String
-
-  private val postGravsearch =
-    SecuredEndpointHandler[(GravsearchQuery, FormatOptions, Option[ProjectIri]), (RenderedResponse, MediaType)](
-      searchEndpoints.postGravsearch,
-      user => { case (query, opts, limitToProject) => searchRestService.gravsearch(query, opts, user, limitToProject) },
-    )
-
-  private val getGravsearch =
-    SecuredEndpointHandler[(GravsearchQuery, FormatOptions, Option[ProjectIri]), (RenderedResponse, MediaType)](
-      searchEndpoints.getGravsearch,
-      user => { case (query, opts, limitToProject) => searchRestService.gravsearch(query, opts, user, limitToProject) },
-    )
-
-  private val postGravsearchCount =
-    SecuredEndpointHandler[(GravsearchQuery, FormatOptions, Option[ProjectIri]), (RenderedResponse, MediaType)](
-      searchEndpoints.postGravsearchCount,
-      user => { case (query, opts, limitToProject) =>
-        searchRestService.gravsearchCount(query, opts, user, limitToProject)
-      },
-    )
-
-  private val getGravsearchCount =
-    SecuredEndpointHandler[(GravsearchQuery, FormatOptions, Option[ProjectIri]), (RenderedResponse, MediaType)](
-      searchEndpoints.getGravsearchCount,
-      user => { case (query, opts, limitToProject) =>
-        searchRestService.gravsearchCount(query, opts, user, limitToProject)
-      },
-    )
-
-  private val getSearchIncomingLinks =
-    SecuredEndpointHandler[(InputIri, Offset, FormatOptions, Option[ProjectIri]), (RenderedResponse, MediaType)](
-      searchEndpoints.getSearchIncomingLinks,
-      user => { case (resourceIri, offset, opts, limitToProject) =>
-        searchRestService.searchIncomingLinks(resourceIri.value, offset, opts, user, limitToProject)
-      },
-    )
-
-  private val getSearchStillImageRepresentations =
-    SecuredEndpointHandler[(InputIri, Offset, FormatOptions, Option[ProjectIri]), (RenderedResponse, MediaType)](
-      searchEndpoints.getSearchStillImageRepresentations,
-      user => { case (resourceIri, offset, opts, limitToProject) =>
-        searchRestService.getSearchStillImageRepresentations(resourceIri.value, offset, opts, user, limitToProject)
-      },
-    )
-
-  private val getSearchStillImageRepresentationsCount =
-    SecuredEndpointHandler[(InputIri, FormatOptions, Option[ProjectIri]), (RenderedResponse, MediaType)](
-      searchEndpoints.getSearchStillImageRepresentationsCount,
-      user => { case (resourceIri, opts, limitToProject) =>
-        searchRestService.getSearchStillImageRepresentationsCount(resourceIri.value, opts, user, limitToProject)
-      },
-    )
-
-  private val getSearchIncomingRegions =
-    SecuredEndpointHandler[(InputIri, Offset, FormatOptions, Option[ProjectIri]), (RenderedResponse, MediaType)](
-      searchEndpoints.getSearchIncomingRegions,
-      user => { case (resourceIri, offset, opts, limitToProject) =>
-        searchRestService.searchIncomingRegions(resourceIri.value, offset, opts, user, limitToProject)
-      },
-    )
-
-  private val getSearchByLabel =
-    SecuredEndpointHandler[
-      (String, FormatOptions, Offset, Option[ProjectIri], Option[InputIri]),
-      (RenderedResponse, MediaType),
-    ](
-      searchEndpoints.getSearchByLabel,
-      user => { case (query, opts, offset, project, resourceClass) =>
-        searchRestService.searchResourcesByLabelV2(query, opts, offset, project, resourceClass, user)
-      },
-    )
-
-  private val getSearchByLabelCount =
-    SecuredEndpointHandler[
-      (String, FormatOptions, Option[ProjectIri], Option[InputIri]),
-      (RenderedResponse, MediaType),
-    ](
-      searchEndpoints.getSearchByLabelCount,
-      _ => { case (query, opts, project, resourceClass) =>
-        searchRestService.searchResourcesByLabelCountV2(query, opts, project, resourceClass)
-      },
-    )
-
-  private val getFullTextSearch =
-    SecuredEndpointHandler[
-      (String, FormatOptions, Offset, Option[ProjectIri], Option[InputIri], Option[InputIri], Boolean),
-      (RenderedResponse, MediaType),
-    ](
-      searchEndpoints.getFullTextSearch,
-      user => { case (query, opts, offset, project, resourceClass, standoffClass, returnFiles) =>
-        searchRestService.fullTextSearch(query, opts, offset, project, resourceClass, standoffClass, returnFiles, user)
-      },
-    )
-
-  private val getFullTextSearchCount =
-    SecuredEndpointHandler[
-      (String, FormatOptions, Option[ProjectIri], Option[InputIri], Option[InputIri]),
-      (RenderedResponse, MediaType),
-    ](
-      searchEndpoints.getFullTextSearchCount,
-      _ => { case (query, opts, project, resourceClass, standoffClass) =>
-        searchRestService.fullTextSearchCount(query, opts, project, resourceClass, standoffClass)
-      },
-    )
-
-  val routes: Seq[Route] =
-    Seq(
-      getFullTextSearch,
-      getFullTextSearchCount,
-      getSearchByLabel,
-      getSearchByLabelCount,
-      postGravsearch,
-      getGravsearch,
-      postGravsearchCount,
-      getGravsearchCount,
-      getSearchIncomingLinks,
-      getSearchStillImageRepresentations,
-      getSearchStillImageRepresentationsCount,
-      getSearchIncomingRegions,
-    )
-      .map(it => mapper.mapSecuredEndpointHandler(it))
-      .map(it => tapirToPekko.toRoute(it))
-}
-
-object SearchApiRoutes {
-  val layer = SearchRestService.layer >+> SearchEndpoints.layer >>> ZLayer.derive[SearchApiRoutes]
-}
-
-final case class SearchRestService(
-  searchResponderV2: SearchResponderV2,
-  renderer: KnoraResponseRenderer,
-  iriConverter: IriConverter,
-  tracing: Tracing,
-) {
-
-  def searchResourcesByLabelV2(
-    query: String,
-    opts: FormatOptions,
-    offset: Offset,
-    project: Option[ProjectIri],
-    limitByResourceClass: Option[InputIri],
-    user: User,
-  ): Task[(RenderedResponse, MediaType)] = for {
-    resourceClass <- ZIO.foreach(limitByResourceClass.map(_.value))(iriConverter.asSmartIri)
-    searchResult <-
-      searchResponderV2.searchResourcesByLabelV2(query, offset.value, project, resourceClass, opts.schema, user)
-    response <- renderer.render(searchResult, opts)
-  } yield response
-
-  def searchResourcesByLabelCountV2(
-    query: String,
-    opts: FormatOptions,
-    project: Option[ProjectIri],
-    limitByResourceClass: Option[InputIri],
-  ): Task[(RenderedResponse, MediaType)] = for {
-    resourceClass <- ZIO.foreach(limitByResourceClass.map(_.value))(iriConverter.asSmartIri)
-    searchResult <-
-      searchResponderV2.searchResourcesByLabelCountV2(query, project, resourceClass)
-    response <- renderer.render(searchResult, opts)
-  } yield response
-
-  def gravsearch(
-    query: String,
-    opts: FormatOptions,
-    user: User,
-    limitToProject: Option[ProjectIri],
-  ): Task[(RenderedResponse, MediaType)] = for {
-    searchResult <- searchResponderV2.gravsearchV2(query, opts.schemaRendering, user, limitToProject)
-    response     <- renderer.render(searchResult, opts)
-  } yield response
-
-  def gravsearchCount(
-    query: String,
-    opts: FormatOptions,
-    user: User,
-    limitToProject: Option[ProjectIri],
-  ): Task[(RenderedResponse, MediaType)] = for {
-    searchResult <- searchResponderV2.gravsearchCountV2(query, user, limitToProject)
-    response     <- renderer.render(searchResult, opts)
-  } yield response
-
-  def searchIncomingLinks(
-    resourceIri: String,
-    offset: Offset,
-    opts: FormatOptions,
-    user: User,
-    limitToProject: Option[ProjectIri],
-  ): Task[(RenderedResponse, MediaType)] =
-    (for {
-      searchResult <-
-        searchResponderV2.searchIncomingLinksV2(resourceIri, offset.value, opts.schemaRendering, user, limitToProject)
-          @@ tracing.aspects.span("query")
-      response <- renderer.render(searchResult, opts) @@ tracing.aspects.span("render")
-      _        <- ZIO.succeed(Sentry.captureMessage("searchIncomingLinks", SentryLevel.INFO))
-    } yield response) @@ tracing.aspects.root(
-      spanName = "searchIncomingLinks",
-      attributes = Attributes
-        .builder()
-        .put("resourceIri", resourceIri)
-        .put("offset", offset.value)
-        .put("limitToProject", limitToProject.map(_.value).getOrElse("None"))
-        .build(),
-    )
-
-  def getSearchStillImageRepresentations(
-    resourceIri: String,
-    offset: Offset,
-    opts: FormatOptions,
-    user: User,
-    limitToProject: Option[ProjectIri],
-  ): Task[(RenderedResponse, MediaType)] =
-    for {
-      response <-
-        tracing.root("searchStillImageRepresentations") {
-          for {
-            result <-
-              searchResponderV2.searchStillImageRepresentationsV2(
-                resourceIri,
-                offset.value,
-                opts.schemaRendering,
-                user,
-                limitToProject,
-              )
-            rr <- renderer.render(result, opts)
-          } yield rr
-        }
-    } yield response
-
-  def getSearchStillImageRepresentationsCount(
-    resourceIri: String,
-    opts: FormatOptions,
-    user: User,
-    limitToProject: Option[ProjectIri],
-  ): Task[(RenderedResponse, MediaType)] =
-    for {
-      response <-
-        tracing.root("searchStillImageRepresentationsCount") {
-          for {
-            result <-
-              searchResponderV2.searchStillImageRepresentationsCountV2(
-                resourceIri,
-                user,
-                limitToProject,
-              )
-            rr <- renderer.render(result, opts)
-          } yield rr
-        }
-    } yield response
-
-  def searchIncomingRegions(
-    resourceIri: String,
-    offset: Offset,
-    opts: FormatOptions,
-    user: User,
-    limitToProject: Option[ProjectIri],
-  ): Task[(RenderedResponse, MediaType)] =
-    for {
-      response <-
-        tracing.root("searchIncomingRegions") {
-          for {
-            searchResult <-
-              searchResponderV2.searchIncomingRegionsV2(
-                resourceIri,
-                offset.value,
-                opts.schemaRendering,
-                user,
-                limitToProject,
-              )
-            response <- renderer.render(searchResult, opts)
-          } yield response
-        }
-    } yield response
-
-  def fullTextSearch(
-    query: RenderedResponse,
-    opts: FormatOptions,
-    offset: Offset,
-    project: Option[ProjectIri],
-    resourceClass: Option[InputIri],
-    standoffClass: Option[InputIri],
-    returnFiles: Boolean,
-    user: User,
-  ): Task[(RenderedResponse, MediaType)] = for {
-    resourceClass <- ZIO.foreach(resourceClass.map(_.value))(iriConverter.asSmartIri)
-    standoffClass <- ZIO.foreach(standoffClass.map(_.value))(iriConverter.asSmartIri)
-    searchResult <- searchResponderV2.fulltextSearchV2(
-                      query,
-                      offset.value,
-                      project,
-                      resourceClass,
-                      standoffClass,
-                      returnFiles,
-                      opts.schemaRendering,
-                      user,
-                    )
-    response <- renderer.render(searchResult, opts)
-  } yield response
-
-  def fullTextSearchCount(
-    query: RenderedResponse,
-    opts: FormatOptions,
-    project: Option[ProjectIri],
-    resourceClass: Option[InputIri],
-    standoffClass: Option[InputIri],
-  ): zio.Task[
-    (_root_.org.knora.webapi.slice.common.api.KnoraResponseRenderer.RenderedResponse, _root_.sttp.model.MediaType),
-  ] = for {
-    resourceClass <- ZIO.foreach(resourceClass.map(_.value))(iriConverter.asSmartIri)
-    standoffClass <- ZIO.foreach(standoffClass.map(_.value))(iriConverter.asSmartIri)
-    searchResult  <- searchResponderV2.fulltextSearchCountV2(query, project, resourceClass, standoffClass)
-    response      <- renderer.render(searchResult, opts)
-  } yield response
-}
-
-object SearchRestService {
-  val layer = ZLayer.derive[SearchRestService]
 }
